"12:45pm. https://boards.4chan.org/a/thread/180340418/baki#p180352063
The post up two from this one that posted the That's My Horse, Ed Edd and Eddy clip was hillarious.

12:50pm.

```
1 -0.5 -0.5
0 0.5 0
0 0 0.5
```

```
1 1 1
0 2 0
0 0 2
```

This is essentially the MDP I was imagining. The inversion does exactly what I thought it would.

The steady state matrix essentially represents transitions and probabilities.

1:10pm. If the diagonals are positive and off diagonals are negative is the matrix always invertible?

1:20pm. https://en.wikipedia.org/wiki/Diagonally_dominant_matrix

Yeah, I remembered this from the Cholesky videos where it was mentioned something about diagonal dominance. I could use the ln+relu to ensure that the inputs to head are always this. Meaning if I could ensure that the diagonals are positive and the off diagonals are negative, I will never have to worry about inverting it. Very simple regularization should suffice.

...Ah, no wait, it meant absolute values.

1:25pm.

```
1 -1
-1 1
```

Ah, so this would be a non-invertible matrix.

This is also telling me something. These kinds of transition probabilities would indicate a circular MDP.

1:40pm.

```
1 -0.5 -0.5
-0.5 1.0 -0.5
-0.5 -0.5 1
```

Yeah, when the probabilities are circular, the matrix is singular.

```
1 -0.75 -0.25
-0.5 1.0 -0.5
-0.25 -0.75 1
```

Also this.

```
1 -0.75 -0.25
-0.7 1.0 -0.3
-0.25 -0.75 1
```

What if I let the off diagonals not sum up to -1?

1:45pm. Yeah, then the matrix instantly become invertible again.

I understand it.

```
1 -0.75 -0.25
-1.4 2.0 -0.6
-1 -1 0
```

The condition for the singularity is that all the elements in all the rows add up to zero.

In a tabular regime I could ensure invertibility with just a tad regularization, but I am not sure what will happen in a realistic setting. I think EMA updates + regularization should give me the margin of safety that I need.
"