"1:55pm. It seems that circularity idea only applies to 3x3 matrices. Since things are like that, let me make sure that the example I thought of with the elegibility traces in the face of stochastic transitions generalizes.
```
1 -1 0 0
0 1 -0.5 -0.5
0 0 0.5 0
0 0 0 0.5
```

This is a relatively simple model to describe.

```
1 1 1 1
0 1 1 1
0 0 2 0
0 0 0 2
```

Yeah, this is exactly what I would expect.

2pm. I am playing with this and I can really see it now. This is just so much more vivid than having a bunch of updates that you know nothing about. There is a whole language packed away in these matrices, more powerful than the probabilistic framework. If a matrix is a model of the environment and its transitions, then its inverse can be used to do credit assignment.

This is what Zap does.

I am completely convinced of this now. The paper is filled with proofs, but what it is doing actually makes sense to me in mental simulations on toy examples. The paper really blew it by not doing this and instead filling it impenetrable math. Nevertheless, I'd be a fool not to take advantage of this while doing RL now that I understand it.

There is one more thing. It concerns MC.

```
1 0   0   0
0 1   0   0
0 0 1/2   0
0 0   0 1/2
```

...It seems that when you paste the matrix it automatically lays it out.

```
1 0 0 0
0 1 0 0
0 0 2 0
0 0 0 2
```

While TD does model based learning when combined with Zap, MC most assuredly does not.

2:15pm. If I want to be doing model based learning, then I need to be doing Zap TD.

2:20pm. While one would want to make policy moves conservative hence the reasoning behind natural gradients, I cannot see a reason why one would want to rescale the reward in the Zap critic. It makes no sense.

2:25pm. I guess this is a case of needing the right tool for the right job. There is still a bit of incompatibility in that I will be doing multiple TD updates and averaging them, but I think that will be fine.

Also, today's experiments confirm to me why the network blew up when I took Aditha's advice and not updated in the last round of the iteration. No gradients at all in the phase of the episode is what I would get in the tabular case.

2:30pm. This is so great. Though I gave up on Zap, it won't be hard to fix that mistake and get the full power of model based RL. The difference between model-based and model-free learning is enormous and the results in their paper certainly demonstrate bridging the gap.

2:35pm. I guess the only other concern is that the critic operates on the top level. It might be plausible that this might not be enough information to do full model based RL.

...Yeah, there might be something to that view.

Consider this, 4 actions, 4 inputs to the head. That might be enough for the actor, but not for the critic. Critic definitely needs to be able to tell states apart.

2:40pm. But even so, the right choice in this hypothetical situation would not necessarily be to let gradients through.

Isn't this the best time to introduce skip connections?

If I have a ten layer net, I can afford to feed the critic the outputs of the top 3 or 4. It is no problem.

2:45pm. Let me see if I can reason it through more what would happen with the gradients..."