"3pm. https://classroom.udacity.com/courses/ud600/lessons/4627968925/concepts/46399385970923
I am reasoning it out. I think there is a decent chance that just KFAC would be enough to stabilize TD on Baird's counter-example. It is a very contrived example. Had the inputs been scaled normally it would not have diverged. But it probably would not converge as fast as with Zap.

3:05pm. Ok, skip connections might work. So might letting the gradients through and leaving the rest to KFAC.

I think now that I understand what Zap is doing, I should just open myself open to providence and just try the Zap optimizer for regular layers. If it good for the critic and for TD learning, it just might be good for regular learning. I mean it is the first thing on the list I should consider trying.

Zap is essentially KFAC's front pass plus eligibility traces. This is not a bad thing to at least attempt trying. I definitely have a reason to think it might work.

And if it does work, it will work very well indeed.

There is just one thing to keep in mind if I do it. So far I haven't been dividing by minibatch size. I used to do this, but then realized that it does nothing on Mnist so I've removed it. Once eligibility traces enter the picture, I should consider adding that back.

3:15pm. ...It will work. From a risk/reward perspective there is no reason to think that Zap will work any worse than KFAC for either feedforward or recurrent layers, but plenty of reason to think why it might do much better. After all, KFAC works with squared error cost functions. While Zap might have evolved from TD linear updates, there is no reason to keep it there. Why not make regular layers sophisticated RL systems?

3:30pm. In the feedforward case Zap is identical to KFAC's front pass. In the recurrent case it generalizes model-based RL with eligibility traces.

I will combine it with plastic RNNs and cover the large gap in sample complexity between machines and humans in one feel swoop.

That is what is important right now. I need to a powerful algorithm to become my power. What I have now and everything I've been working on for the past 3.5 years is still solidly in the useless stage. Once I have that power I will have much more leeway to do whatever I want, whether that be making Spiral better or making money using RL. I will have the leeway to consider resuming Simulacrum. I could use the plastic RNNs to generate art and music.

It does not have to be simply money, but I need to move away from zero. I need that power.

That Maze task from the Backpropamine paper is something I need to master in 100s or 1000s of iterations, not tens of thousands. If Zap and plastic RNNs can give that victory to me, I will have a real accomplishment."