I think I will step away from the screen for a while. My mind has been generating ideas in a feverish, high intensity state and I need to come down from that high. It started during the night so I haven't gotten good sleep. I think I will call it a day here.
I need to relax.

Tomorrow, I think I will do the reading that I wanted to do today and then...well, I think I will just test the Zap idea in the inner layer on the char RNN task. Rather than let this weight on me, I should just test my hypothesis and then move on to the next thing.

That is the way of the scientist, mad or not.