I had a good night's rest and my head is cleared up. This might not seem like much, but I was heated the entire night and day yesterday. My mind feels like an oasis in comparison to yesterday's.
Yeah, I understand what I am undertaking by suggesting that Zap would eliminate the need for BPTT. If that turns out to be true, then Zap and plasticity will turn my neural net agents into living systems.

I concluded that Zap would be good for the critic. That I know for certain based on what it does in the linear tabular regume. But ignoring its origin and expanding on its purpose would be a radical move.

I remember thinking about how I would do backprop through time past its truncation horizon and essentially coming to the conclussion that some kind of model would be required to make an approximate update.

Zap's inverse steady state matrix is exactly that model. In the tabular case it does optimal credit assignment and there is no think it would not generalize to the general linear case.

So as radical as it is, this experiment is simply going to have to work.

I am going to do another redesign to make sure that everything flows smoothly. When I do the AC, I want to merge the actor and the critic into a single multiply, but since I've decided to do not want KFAC's backward pass in the critic that will create trouble.

Moreover, at some point I might have to do very large multiplies so it might be worth it to split KFAC and have it work on separate submatrices. A little flexibility here so I can do this properly should be good.

I am really happy that I laid the groundwork for passing future states to the past in the backward pass everywhere. I said I would eventually test Zap in the inner layers even though I had no concrete reason to do so at the time. I am glad I did not change my mind and gotten rid of the baggage.

It definitely does feel that my true talent might lie in algorithm selection more than anything else. Well, that is only if this actually turns out to be like I want it to.

I'll do the Zap experiment on the `zap` branch. The `master` won't be updated for a bit until then.