```             inl update k data =                 match d with                 | {$k={covariance k}} ->                     update_covariance learning_rate data covariance s                     k := k() + out.span_outer                 | _ -> ()             update .front (primal data)             update .back out ```
In addition to covariance I will just check for steady state.

```
            inl update k data =
                match d with
                | {$k={steady_state k}} ->
                    update_steady_state learning_rate data steady_state s
                    k := k() + out.span_outer
                | {$k={covariance k}} ->
                    update_covariance learning_rate data covariance s
                    k := k() + out.span_outer
                | _ -> ()
```

Let me do it like this.

Easy as pie. Actually I won't make the changes just for the sake of AC. I'll forget that for the time being. Actually keeping it together for the matrix multiply and separate during optimization is just too much of a chore. I'll go with what I have.

Doing a little extra work won't kill the GPU. I always obsess about trivial bullshit like that. It is a good exercise, but I need to focus on what is important. I'll leave optimizing it 100% for the future. What I need to get right are the main, global optimizations. The rest will take care of itself.

Let me bring in the `update_steady_state` from the old module.