"2:25pm.
```
    inl regularized_lu_inverse {epsilon from to} =
        inb from = s.CudaFun.map {mapi=dampen epsilon} from |> CudaAux.temporary
        inl {ipiv info} = getrf' s from
        inb ipiv = CudaAux.temporary ipiv
        handle_error s { info pos = "U(x,x) = 0 where x = %d."}
        set_to_identity to
        handle_error s {info = s.CudaSolve.getrs' .nT from ipiv to}
```

This is how it should go.

```
    inl cholesky_inverse s = function
        | {from to} ->
            symm_copy s .Lower {from to}
            handle_error s {
                info = potrf' s .Lower to
                pos = "The leading minor of order %d is not positive definite."
                }
            cholesky_body to
        | {from} | from ->
            inl C_sqrt = potrf s .Lower from
            cholesky_body C_sqrt
            C_sqrt
```

The way these functions are designed is just so weird. Why isn't the `{from}` case just calling `{from to}`? I botched the design here, but I do not feel like redoing it right now. What is important now is to get Zap operational. Let me plug the new function and then I will be on my way."