"5:25pm.
```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:07.2451517 for: iteration 0
Training: 5.47455455297056
```

I switch back to the Cholesky update and am still getting this for some reason. It is strange that the error is not moving at all.

Let me turn off KFAC.

```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:06.9199778 for: iteration 0
Training: 5.47455455297056
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:06.9139552 for: iteration 1
Training: 5.47455455297056
The learning rate is 2 ** -13
```

It is not updating. That is the problem.

Considering KFAC is not active, this thing got damn slow all of a sudden.

5:30pm. The backward functions are running.

I have no idea at all what the problem is.

Let me bring in the previous Learning module and paste it over what I've made now. I think the error is in matmult for some reason, but I cannot imagine what would cause the updates to not happen like this."