"5:45pm.
```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:07.2266980 for: iteration 0
Training: 2.26654464491944
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:07.1562888 for: iteration 1
Training: 1.83886282828719
The learning rate is 2 ** -13
Starting timing for: iteration 2
```

Ok, now let me try the steady state thing again.

```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:08.3593421 for: iteration 0
Training: 3.05058120277382
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:08.1667979 for: iteration 1
Training: 2.40370797936339
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:08.3187217 for: iteration 2
Training: 2.28284297177486
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:08.2479550 for: iteration 3
Training: 2.22112259485841
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:08.3735257 for: iteration 4
Training: 2.18089296220227
```

This is about SGD level of performance.

5:50pm. How sure am I that I am using the matrix the right way around?

...Makes no real difference.

Let me make sure that it is actually doing the subtraction of states.

Also is this really SGD levels of performance? What if I reduce the number of timesteps to something like 8?"