"7:15pm. Let me just say that I am disappointed by how today went. The greedy idea came up empty, and it does not feel like I did much of note.
I started the day extremely cleanheaded and refreshed and ended up essentially slacking. Well, I earned it after a day like yesterday, but I must endeavor to do better from here. In the end it turned out that Zap is LSTD and it was discovered back in 1996. And the great intuitive leap I had that made me discover that LSTD is essentially doing model based learning was known in 1998.

But nevertheless, I pushed my mind to its limit to discover that and the reason why I even found the LSTD reference today was because of that. It led me to this.

Looking at the level of sophistication in the new Zap paper is making me realize that the expertise needed to develop new optimization algorithms is not a laughing matter at all. It all leads back to the conclusion that I need the help in form of AI.

Ignoring the hidden layer RNN gamble, the rest of the reasoning I did yesterday was sound. KFAC for the actor, LSTD for the critic, and KFAC for the hidden layers. Just plasticity itself will have to be enough to get me where I want.

But I feel like I'd missed an opportunity. If it were possible to unlock the full power of unsupervised model based RL learning in the hidden layers, I think that along with plasticity would have set the Singularity happening in stone in the 20s.

A lot stronger effort will be needed. Damn, the dream yesterday was just so strong. One day I will reach it.

But since it is possible to do in the final layer, that means there probably is more to discover in the inner ones as well. It was definitely an real eye opener of exactly what is possible with smart optimizers.

7:40pm. Let me call it a day for real here.

Before I got distracted by all of this, my intent was to do the Maze task. I will intensify my effort starting from here.

As soon as I am done investigating the performance regression, I will get the poker game and the RL layers back online and start work on this.

I intent to be done with it in less than a week, but given how long testing will take, I except this to last a while. In think in December I will have time to seriously dive into poker with the plastic RNN that does reversible learning in the inner loop.

All these bumps will even out in the long run.

7:50pm. There is one goal I need to add to the mix. It is interesting that LSTD does not work any better than MC when I tried it. I am going to have to get to the bottom of that and the Maze game will be the ideal platform to do this.

Also I am going to test whether LSTD works better than KFAC for the critic in a deep net. This is something I need to make sure of. It might turn out to be the case that the inner layers are conspiring to make the critic give good value estimates. This is something I need to keep in mind.

Supposing that LSTD does not give any benefit for the critic in a plastic RNN, then that will be powerful evidence to turn away from optimizers and towards models. Otherwise, I will expect that there will be good discoveries to be expected beyond KFAC for the inner layers."