"9:25am. https://arxiv.org/abs/1811.02553 Are Deep Policy Gradient Algorithms Truly Policy Gradient Algorithms?
This was just posted today. I'll take some time to read this. Also, there was a paper that I skimmed that I did not mention yesterday that bothered me.

https://arxiv.org/abs/1705.07461
Shallow Updates for Deep Reinforcement Learning

Wow, does this one bother me. So using a running average with LSTD does not work in a deep setting?

10am. "We finish this Section with an interesting observation. While the LS solution improves the performance
of the DRL agents, we found that the LS solution weights are very close to the baseline DQN
solution. See Appendix D, for the full results. Moreover, the distance was inversely proportional to
the performance of the solution. That is, the FQI solution that performed the best, was the closest (in
`2 norm) to the DQN solution, and vice versa. There were orders of magnitude differences between
the norms of solutions that performed well and those that did not. Similar results, i.e., that large-batch
solutions find solutions that are close to the baseline, have been reported in (Keskar et al., 2016). We
further compare our results with the findings of Keskar et al. in the section to follow.
"

10am. There is something I need to consider. Given how often I am using matrix inversion, I think I would be better off by factoring and solving equations explicitly.

10:05am. I am going to ask the author about using a running average to update A and then applying the algorithm on every iteration instead. This just bothers me.

After that I'll read the first paper."