"10:45am. Note, FQI is exactly KFAC. I am going to have to check it out. It seems to perform better than LSTD in DQN.
11:05am. Done reading PG-PG paper. It is interesting, though not directly applicable to me as I do not intend to use surrogate objectives and PPO.

Let me go back to the shallow updates for RL so I can get the FQI reference. I have a strong reason to consider it - Cholesky factorization is much faster than LU for inversion. If it turns out that FQI performs as well as LSTD in practice, I'll just go entirely to it in that case.

For a further speed up, now that I've centralized the weights, I think I will in fact implement direct solving. The great thing is that `getrs` does it all in place. That was an annoyance to me, but it could be huge opportunity.

11:20am. http://ml.informatik.uni-freiburg.de/former/_media/publications/rieecml05.pdf
Neural Fitted Q Iteration - First Experiences with a Data Efficient Neural Reinforcement Learning Method

http://www.jmlr.org/papers/volume6/ernst05a/ernst05a.pdf
Tree-Based Batch Mode Reinforcement Learning

Let me read the FQI papers and then I will take a break.

11:35pm. Interesting that NFQ would use RProp. This is like SGD except with only the sign of the gradient considered. I tried it back in 2015.

11:40am.

```
The basic idea underlying NFQ is the following: Instead of updating the neural
value function on-line (which leads to the problems described in the previous
section), the update is performed off-line considering an entire set of transition
experiences. Experiences are collected in triples of the form (s, a, s
) by interacting
with the (real or simulated) system1. Here, s is the original state, a is the
chosen action and s is the resulting state. The set of experiences is called the
sample set D.
```

I keep wondering why the b * b^T is nowhere to be found, but a flash of insight is coming to me. Remember the Fast Weights paper said that multiplying the A by any vector is equivalent to multiplying the whole trajectory by it?

I need some more info.

11:50am. https://youtu.be/AohtWrJwvgg

This is a good lecture. Let me finish it and I will take a break.

12:30pm. I am starting to have an idea.

https://groups.google.com/forum/#!topic/rl-list/tII1qZfXpOA
LSPI vs FQI

1:05pm. I made a great insight. With the LSTD update, I could do policy gradients by passing the rewards directly to the actor without either MC or TD! I found it, I finally found the thing!

It is possible to avoid having to do MC and TD both.

2:45pm. Done with breakfast.

Actually, I could not do the PG updates with the Zap update. It works for TD because it is linear, but for PG, I'd have to somehow compensate for the gradients as well too. `b * (b - b')` is just not the right update. It is missing something. The same goes for LSTD in a deep net. The reason why it works worse than KFAC is because of lower layers. The lower layers and the head are being optimized with mutually exclusive objective functions.

I know that SGD can deal with it, but this is an exception. Normally, backprop would compensate, but the LSTD update wrenches the layers away. That is why I get only noise on the char RNN when I'd tried it.

The middle layers and PG would need their own update.

2:55pm. Let me make a commit with these entries.

I am just so fascinated with this now. I want to do more research on this.

I am struck by inspiration. I'll probably get nowhere with this, but I just do not feel like programming at the moment. It is so humiliating. It feels like these is a wealth to be mined here, but I cannot draw even a single insight with my own power.

If I could make those plastic RNNs and the optimizers for them, I could easily tap into enough power to make a beeline towards the Singularity.

I mean, just on its own, that would not be enough - currently, NNs lack the modularity needed to overcome catastrophic forgetting and there is also the matter of exploration, but having those optimizers and the plastic RNNs would be like half the story done already. LSTD is just so frustrating. If only a tiny bit more could be drawn from that thread it would be possible to get everything.

3:05pm. But well, maybe using a LSTD critic disconnected from the main net is the optimal form?

When I get serious, I think I will just go for this.

Right now, let me do some more study."