I studied quite intensely today. On the side, 'MC vs TD vs MC vs TD...' battle keeps repeating in my head. Do I let the gradients flow from the critic? Do I train it using MC and the actor using TD?
I've been having enormous difficulty deciding.

But I've come to a decision. While it is true that MC works a tiny bit better TD, what I am going to do is try and maximize the potential of the metalearning architectures I am trying out.

If the plastic RNNs turn out to generalize well, there is a possibility that they might hijack the TD function and learn to do the right thing in advance.

This is something that would be impossible if I block the critic gradients or use MC. So I am going to have to use KFAC and just let the gradients flow.

That is the simplest solution and it is not that much worse than MC on the toy task that I have to avoid it.

https://arxiv.org/abs/1705.07461
Shallow Updates for Deep Reinforcement Learning

This paper is the key that decides whether I should go with LSTD or KFAC. KFAC without a doubt. They are using that update and it works better than LSTD. This is convincing proof that LSTD is made for the linear case and will give me issues with deep nets.

Maybe at some point Aditha Devray might figure out how to adapt Zap to deep nets, but for now just forget it. It will take miracle. As far the events move from here, that should be enough to understand how far that approach will lead. Instead of bashing my head against this and wasting any more time, I should redouble my efforts on metalearning architectures.

Track down the performance regression, optimize how KFAC is done, bring RL back online and then move to the Maze task.

Then freedom will be in my grasp.

In the past few days I had the divine taste of model based RL, but it is fake. Must like self improvement, the real model based RL is in a different direction.

It will take a rocky path to get there, but I will persevere.

I will take it easy tomorrow. I'll just focus on finding the performance regression and speeding up KFAC. I'll move to doing a decomposition on each step.

Also, I should have done this Zap stuff on a different branch just like I said I would, but I forgot about it right after saying that I would do it. I'll back it up tomorrow.

Fundamentally, I believe in TD learning.

And I must believe that it is necessary for metalearning. The Shallow Updates paper is a crucial hint pointing in that direction.

To make good decisions, It is important to understand that there are different kinds of errors. On paper, LSTD should vastly outperform KFAC, but that is not happening in deep nets. Why?

It is more than just random shifting of the basis that occurs at the top due to the influence of lower layers.

Probably, this is due to the effect of generalization - even in standard feedforward nets, the hierarchy ensures a degree of metalearning. This causes the model in the head to be highly inaccurate.

The same could be said for KFAC in fitter Q iteration, but it seems the kinds of errors caused by basis shifting is something it can cope with. LSTD managed to do well where it matches KFAC and that is it.

From my current vantage point, I have absolutely no idea how I am going to learn models needed to bring about full AGI, but one this is certain that future devices with terrabytes and petabytes of RAM specialized for the kinds of computations required by NNs will be a great help.

https://www.youtube.com/watch?v=8a7wBLg5Q8U
Keynote Talk: Model Based Machine Learning

I'll watch this talk by Chris Bishop and then call it a day.