"9:40am. Well, it was an interesting paper. I guess I'll post it on the RL sub.
10:10am. "It is interesting how they instead of using a target network, they store the trace directly in the replay buffer. Little algorithmic tricks like this are underappreciated.

From the paper, it seems DQN's benefit a lot more from eligibility traces than A3C for some reason.

I have to admit, for a long time ever since I learned about eligibility traces I thought of them as some kind of algorithmic hack, but just very recently I started to see them as the missing link between higher order optimization (like what KFAC is doing) and model based RL.

I do not mean the direct way like it is done in this paper. Rather if one takes Zap/LSTD's steady state matrix and observes what its inverse is doing on simple linear tabular examples, one will see that the eligibility traces are right there in the steady state matrix's inverse. Model based RL has eligibility traces directly in the model and they are the magic ingredient that allows it to do optimal credit assignment.

This has had an impact on me as it is the first truly realistic example of model based RL that I'd seen so far.

It is a pity that LSTD does not [work better](https://arxiv.org/abs/1705.07461) than KFAC (FQI) in a deep setting. I hypothesize that the reason for that is because doing LSTD in just the top layer has the effect of fundamentally changing the objective for which it optimizes in a way that is invisible to lower layers and that is not compensated by SGD.

If one could derive an update similar to LSTD's that compensates for the impact of the nonlinearities in the hidden layers, then it might be possible to get the full benefit of model based RL in deep networks. Between that and plastic RNNs, that should be enough to cross the 3x order of magnitude gap needed for RL to become really useful."

I wrote this. Agh, it seems today's morning will be another time waster.

You know, let me track down the regression and then I'll have that break. Seriously..."