"12:15pm. I'll remove `weight_ss` but otherwise keep the rest of the stuff in. No need to get rid of it just yet. I'll leave the speed optimizations for later.
Right now, as the first thing...How about I enable KFAC in the linear layers?

12:25pm. Actually, I am looking at the experimental results in the appendix and eligibility traces are not a clear win everywhere.

But the formulation is extremely simple and elegant.

```
R = r + gamma * (lambda R' + (1 - lambda) V')
```

Rather than having a MC agent or a TD(0) agent, I should just implement this and call it a job well. That should be the way to go. Then I will be able to smoothly interpolate between TD and MC.

One thing I am thinking about is making lambda an optimizable parameter. I could make it dependent on the inputs. This is the way of plastic RNNs...

12:35pm. Hmmmmm...well, I'll keep the idea in mind.

```
**Edit**: `R = r + gamma * (lambda R' + (1 - lambda) V')`

It might be worth considering making `lambda` a parameter dependent on the inputs. That should be a part of some kind of metalearning approach and would bring the network closer to doing model based RL. Is there any reason why that should not work?
```

I decided to edit my post and added this idea.

1:05pm. I know I am supposed to be working, but what I just wrote above is an incredible idea! It is it! This is the missing link between model based and model free RL! Controllable eligibility traces! It is going to be truly difficult to make a LSTD approximation for deep layers, but something like is extremely reasonable in comparison.

I once saw a different formulation of eligibility traces in Adam White's thesis. There rather than being a mixture of returns at different timescale, the lambda was taken to mean a probability of episode's termination. So there is a piece of evidence in pointing that lambda in fact should not be a hyperparameter of even a parameters, but something that is calculated on the fly.

I'd of course pass it through a sigmoid. Or maybe a rescaled tanh.

Eligibility traces are the difference between model based and model free RL. That is something I am certain of.

Hence allowing the network to control lambda would allow it in theory to learn optimal credit assignment. Conceptually this idea would solve, so so many problems I have right now.

1:20pm. With this the MC vs TD battle has been resolved, for good this time. The result of the war is victory for both sides. I love this kind of hope.

Look, I am sorry. I am going to have to delay my entry into work once more. Right now I feel like taking the time to digest the new idea. What I will do is go through that eligibility trace reference in the Backpropamine paper. I still haven't gotten to that. I only skimmed that paper. Now that I have this concept in mind, this gives eligibility traces an entirely different light."