Enough. I've calmed down a bit. There is no point in studying any longer.
Let me start. Sure this is not as good as the ideal second order optimizer which does exactly model based RL, but that optimizer is highly likely to not exist. Hence for deep nets, having TD with controllable eligibility traces is probably the pinnacle of design.

I mean it is no wonder I skipped over eligibility traces. Everywhere I looked they were like 2-4 steps optimal. That is a fucking joke. A hack if I ever saw one.

I am not sure who is stupider - me or the entire field of RL for overlooking the fact that they can be made learnable. Control the bias variance tradeoff? What a load of bull.

Eligibility traces are to be used to control the length of a learning episode. In real world tasks that length will never be something static that you can just set to. You'd always want to do it granularly to get close to the ideal model based update. Different predictions will have different terminating conditions.

All this started a few days ago when one night I started thinking about eligibility traces and how I might have overlooked some important aspect of them. Just what devilish luck do I have that this paper would be posted at this juncture? Truly the heavens must be telling me to go forward with this.

This sort of thing happens sometimes.

You know, now that I do not need to hope for some miraculous optimizer to come along in order to solve this problem, I think I will start things off by removing the steady state stuff from the ML library.

I can do it. I can do it with this. As it always has my desire will see me through. That is the main thing that I have in abundance.