"4:10pm.
```
    inl td {r gamma lambda R' V' V} =
        inm R = r + gammma * (lambda * R' + (one - lambda) * V')
        {R' = R; out = R - V}
```

This should do it...Ah, I forgot `sqr`.

```
    inl td {r gamma lambda R' V' V} =
        inm R = r + gammma * (lambda * R' + (one - lambda) * V')
        {R' = R; out = sqr(R - V)}
```

Now this reminds me, I am going to have to set the adjoint of `out` to 1 on the backward pass. The standard map won't suffice. I am going to make a custom function for costs.

...But otherwise this should be fine.

This is `td` simplified to the limit and quite efficient to boot.

```
    inl td {r discount_factor eligibility_decay R' V' V} =
        inm R = r + discount_factor * (eligibility_decay * R' + (one - eligibility_decay) * V')
        {R' = R; out = sqr(R - V)}
```

Let me make it readable rather than using Greek letters.

4:25pm.

```
    inl sample_body prob s =
        inl b, a as dim = prob.dim
        inl boundary = s.CudaRandom.create {dst=.Uniform} {elem_type=float; dim=b}
```

The sampling functions are a bit nasty. They are too much of special cases. Note that I missed an opportunity to make the this variable temporary.

4:30pm. No, I fundamentally do not feel like optimizing `sampling_pg`. It is ugly, but let it stay that way.

Let me get the td cost operational and I will bring the layers back on as well.

Well, it will be just one with a single matrix multiply and 3 outputs - `{actions_probs eligibility_decay V}`. I'll have to make it so that only `action_probs` get KFACed on the backward pass. I'll have to modify both `matmult_stream` and `kfac` to enable this.

I am going to have to need some kind of iterative view for this. Or maybe I should just use split.

Actually, how about `split_into`? That should be ideal.

4:40pm. ...I'll think about it later. Let me make a cost function that is based of `init`.

4:45pm. Ah, I get it. I get it how I can do this elegantly. I won't even let `out` have an adjoint. Rather I will make a special function called `as_cost`."